+++ lumin/nn/models/blocks/head.py
@@ -1,5 +1,5 @@
 import numpy as np
-from typing import Dict, Optional, Callable
+from typing import Dict, Optional, Callable, List
 from glob import glob
 from collections import OrderedDict
 from pathlib import Path
@@ -17,10 +17,13 @@ from .abs_block import AbsBlock
 
 
 class AbsHead(AbsBlock):
-    def __init__(self, n_cont_in:int, cat_embedder:Optional[CatEmbedder]=None, 
+    def __init__(self, cont_feats:List[str], cat_embedder:Optional[CatEmbedder]=None, 
                  lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init, freeze:bool=False):
         super().__init__(lookup_init=lookup_init, freeze=freeze)
-        self.n_cont_in,self.cat_embedder = n_cont_in,cat_embedder
+        self.cont_feats,self.cat_embedder = cont_feats,cat_embedder
+        self.n_cont_in = len(cont_feats)
+
+    def _map_outputs(self) -> Dict[str,List[int]]: pass
 
 
 class CatEmbHead(AbsHead):
@@ -29,11 +32,11 @@ class CatEmbHead(AbsHead):
     Provides inputs for continuous features and embedding matrices for categorical inputs, and uses a dense layer to upscale to width of network body.
     Designed to be passed as a 'head' to :class:ModelBuilder.
     Supports batch normalisation and dropout (at separate rates for pre-dense continuous, pre-dense categorical embeddings, and post dense).
-    Continuous features are expected to be the first n_cont_in columns of input tensors and categorical features the remaining columns.
+    Continuous features are expected to be the first len(cont_feats) columns of input tensors and categorical features the remaining columns.
     Embedding arguments for categorical features are set using a :class:CatEmbedder.
 
     Arguments:
-        n_cont_in: number of continuous inputs to expect
+        cont_feats: list of names of continuous input features
         do_cont: if not None will add a dropout layer with dropout rate do acting on the continuous inputs prior to concatination wih the categorical embeddings
         do_cat: if not None will add a dropout layer with dropout rate do acting on the categorical embeddings prior to concatination wih the continuous inputs
         cat_embedder: :class:CatEmbedder providing details of how to embed categorical inputs
@@ -41,15 +44,15 @@ class CatEmbHead(AbsHead):
         freeze: whether to start with module parameters set to untrainable
 
     Examples::
-        >>> head = CatEmbHead(n_cont_in=30)
-        >>> head = CatEmbHead(n_cont_in=25, cat_embedder=CatEmbedder.from_fy(train_fy))
-        >>> head = CatEmbHead(n_cont_in=25, cat_embedder=CatEmbedder.from_fy(train_fy), do_cont=0.1, do_cat=0.05)
-        >>> head = CatEmbHead(n_cont_in=25, cat_embedder=CatEmbedder.from_fy(train_fy), lookup_init=lookup_uniform_init)
+        >>> head = CatEmbHead(cont_feats=cont_feats)
+        >>> head = CatEmbHead(cont_feats=cont_feats, cat_embedder=CatEmbedder.from_fy(train_fy))
+        >>> head = CatEmbHead(cont_feats=cont_feats, cat_embedder=CatEmbedder.from_fy(train_fy), do_cont=0.1, do_cat=0.05)
+        >>> head = CatEmbHead(cont_feats=cont_feats, cat_embedder=CatEmbedder.from_fy(train_fy), lookup_init=lookup_uniform_init)
     '''
 
-    def __init__(self, n_cont_in:int, do_cont:float=0, do_cat:float=0, cat_embedder:Optional[CatEmbedder]=None, 
+    def __init__(self, cont_feats:List[str], do_cont:float=0, do_cat:float=0, cat_embedder:Optional[CatEmbedder]=None, 
                  lookup_init:Callable[[str,Optional[int],Optional[int]],Callable[[Tensor],None]]=lookup_normal_init, freeze:bool=False):
-        super().__init__(n_cont_in=n_cont_in, cat_embedder=cat_embedder, lookup_init=lookup_init, freeze=freeze)
+        super().__init__(cont_feats=cont_feats, cat_embedder=cat_embedder, lookup_init=lookup_init, freeze=freeze)
         self.do_cont,self.do_cat, = do_cont,do_cat
         if self.cat_embedder is None: self.cat_embedder = CatEmbedder([], [])
         if self.cat_embedder.n_cat_in > 0:
@@ -59,6 +62,15 @@ class CatEmbHead(AbsHead):
         self.n_out = self.n_cont_in if self.cat_embedder.n_cat_in == 0 else self.n_cont_in+np.sum(self.cat_embedder.emb_szs)
         if self.do_cont  > 0: self.cont_in_do = nn.Dropout(self.do_cont)
         if self.freeze: self.freeze_layers()
+        self._map_outputs()
+
+    def _map_outputs(self) -> Dict[str,List[int]]:
+        self.feat_map = {}
+        for i, f in enumerate(self.cont_feats): self.feat_map[f] = [i]
+        offset = self.n_cont_in
+        for f, _, sz in self.cat_embedder:
+            self.feat_map[f] = list(range(offset, offset+sz))
+            offset += sz
         
     def forward(self, x_in:Tensor) -> Tensor:
         if self.cat_embedder.n_cat_in > 0:
@@ -68,7 +80,7 @@ class CatEmbHead(AbsHead):
         if self.n_cont_in > 0:
             x_cont = x_in[:,:self.n_cont_in]
             if self.do_cont > 0: x_cont = self.cont_in_do(x_cont) 
-            x = torch.cat((x, x_cont), dim=1) if self.cat_embedder.n_cat_in > 0 else x_cont
+            x = torch.cat((x_cont, x), dim=1) if self.cat_embedder.n_cat_in > 0 else x_cont
         return x
     
     def _load_embeds(self, path:Optional[Path]=None) -> None: